Report editor review

+++ tasks/28.js
@@ -35,32 +35,47 @@ var login = function(email, password){
 
 var removeProject = function(projectName, isIdempotent){
     /*jshint camelcase:false */
-    var token = credentials.access_token;
-    $.util.log('Deleting project ' + projectName);
     var defered = Q.defer();
-    $28.deleteProject(projectName, token).then(function(){
-        $.util.log('Project deleted.');
-        defered.resolve(credentials);
-    }).catch(function(error){
-        if(!isIdempotent) {
-            defered.reject(error);
-        } else {
+    if(!Config.isOnProduction) {
+        var token = credentials.access_token;
+        $.util.log('Deleting project ' + projectName);
+        $28.deleteProject(projectName, token).then(function () {
+            $.util.log('Project deleted.');
             defered.resolve(credentials);
-        }
-    });
+        }).catch(function (error) {
+            if (!isIdempotent) {
+                defered.reject(error);
+            } else {
+                defered.resolve(credentials);
+            }
+        });
+    } else {
+        $.util.log('Skipping project deletion for production: ' + projectName);
+        defered.resolve(credentials);
+    }
     return defered.promise;
 };
 
 var createProject = function(projectName){
     /*jshint camelcase:false */
-    var token = credentials.access_token;
-    $.util.log('Creating project ' + projectName);
-    return $28.createProject(projectName, token).then(function(response) {
-        $.util.log('Project created.');
-        /*jshint camelcase:false */
-        credentials.project_tokens['project_' + projectName] = response.body.projectToken;
-        return credentials;
-    });
+    var defered = Q.defer();
+    if(!Config.isOnProduction) {
+        var token = credentials.access_token;
+        $.util.log('Creating project ' + projectName);
+        $28.createProject(projectName, token).then(function (response) {
+            $.util.log('Project  ' + projectName + ' created.');
+            /*jshint camelcase:false */
+            credentials.project_tokens['project_' + projectName] = response.body.projectToken;
+            defered.resolve(credentials);
+        }).catch(function (error) {
+            $.util.log('Project creation failed: ' + error);
+            defered.reject(error);
+        });
+    } else {
+        $.util.log('Skipping project creation on production: ' + projectName);
+        defered.resolve(credentials);
+    }
+    return defered.promise;
 };
 
 var upload = function(projectName){
@@ -127,14 +142,26 @@ var runQueries = function(projectName, runQueries) {
 };
 
 var createDatasource = function(projectName, datasource){
-    $.util.log('Creating datasource ' + datasource.name);
-    var difault = datasource.default ? datasource.default : false;
-    /*jshint camelcase:false */
-    var projectToken = credentials.project_tokens['project_' + projectName];
-    return $28.createDatasource(projectName, datasource.category, datasource.name, projectToken, difault, JSON.stringify(datasource.credentials)).then(function(){
-        $.util.log(datasource.name + ' created');
-        return credentials;
-    });
+    var defered = Q.defer();
+    if(!Config.isOnProduction) {
+        $.util.log('Creating datasource ' + datasource.name);
+        var difault = datasource.default ? datasource.default : false;
+        /*jshint camelcase:false */
+        var projectToken = credentials.project_tokens['project_' + projectName];
+        $28.createDatasource(projectName, datasource.category, datasource.name, projectToken, difault, JSON.stringify(datasource.credentials))
+        .then(function(){
+            $.util.log(datasource.name + ' created');
+            defered.resolve(credentials);
+        })
+        .catch(function (error) {
+            $.util.log('datasource creation failed: ' + error);
+                defered.reject(error);
+        });
+    } else {
+        $.util.log('Skipping data source creation on production: ' + datasource.name);
+        defered.resolve(credentials);
+    }
+    return defered.promise;
 };
 
 gulp.task('28:login', function(){
